Update switch chain example, update README
